refactor: make core free of display so a GUI can reuse it
Move the priced usage and analysis summary aggregation out of the display
layer into core, relocate the shared per-provider totals container into
models, extract the usage pricing-then-scan pipeline into the library, run
the quota one-shot command from the CLI, and install the terminal panic
hook from the binary instead of core logging. No module under core now
depends on display.
