+++ src/quota/mod.rs
@@ -11,7 +11,6 @@ pub mod claude;
 pub mod codex_session;
 pub mod copilot;
 pub mod cursor;
-pub mod fetch;
 pub mod http;
 pub mod provider;
 pub mod refresh;
